pyproject.toml
[project]
name = "effective-mobile-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.3",
]
